Add an initial Docker configuration
This change adds the initial Docker setup, providing an environment
where the tests can be run, regardless of the user's operating system.
It provides three services; one for PHP, one for MySQL, and one for
PostgreSQL. The PHP environment has all of the dependencies and
extensions required to interact with MySQL, PostgreSQL, and SQLite,
along with any required binaries. It also updates PHPUnit's
configuration, where required, to sync with that configuration.

The change also removes the Vagrantfile as it's no longer required.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>