Revise the Docker Compose configuration
This removes the Composer setup from the startup script, replacing it
with mounting the files from the local filesystem as volume mounts. It
also exposes the database ports so that they can be connected to
locally, should the need arise, using database clients. It adds health
checks to help ensure a robust service configuration. And, it loads SQL
scripts to initialise the MySQL and PostgreSQL databases.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>